examples/python_vulnerable_project/requirements.txt
Flask==1.1.2
Werkzeug==0.16.1
PyYAML==5.1
Requests==2.20.0
Jinja2==2.10.1
itsdangerous==0.24
MarkupSafe==1.0